refactor(node): organize the adapter by folder — src/bin/ and src/mutation/ (#246)
Folder-per-domain instead of flat files:

- `src/bin/index.ts` — the launcher (was `bin.ts`); package `bin` → `dist/bin/index.js`.
- `src/mutation/index.ts` — exposes `mutationCLI`, the pure async orchestrator, alongside
  its one-function-per-file helpers (`parse-args`, `run-stryker`, `to-normalized`,
  `normalize-status`).
- `src/mutation/main.ts` — the executable the rust binary spawns → `dist/mutation/main.js`.
  Kept separate from `index.ts` so the pure function and the process-side-effect stay
  independently covered (a single file with an `isMain` guard would leave the run-block
  uncovered).

The launcher points at `../mutation/main.js`; the rust test helper and the
`--ts-mutation-adapter` docs follow the new path. All internal — no consumer-visible
change beyond the `bin` resolving to `dist/bin/index.js`.
