+++ packages/node/CHANGELOG.md
@@ -7,17 +7,18 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
-- **TypeScript mutation engine adapter** (#246, part of the #239 epic). `src/index.ts` exposes
-  `mutationCLI` — the async orchestrator — over one-function-per-file helpers nested under
-  `src/mutation/` (`parse-args`, `run-stryker`, `to-normalized`, `normalize-status`); the thin
-  `mutation-cli.ts` shim is the executable that runs it. The adapter drives Stryker through its
+- **TypeScript mutation engine adapter** (#246, part of the #239 epic). Organized by folder:
+  `src/mutation/index.ts` exposes `mutationCLI` — the async orchestrator — over one-function-per-file
+  helpers alongside it (`parse-args`, `run-stryker`, `to-normalized`, `normalize-status`), and
+  `src/mutation/main.ts` is the executable that runs it. The adapter drives Stryker through its
   **Node API** (`new Stryker(opts).runMutationTest()`) and maps the structured `MutantResult[]` onto
   the normalized schema the Rust core gates on, selecting the bundled `@stryker-mutator/vitest-runner`
   by resolved path so the unit-scoped runner runs (#240) and reading results in-process (written to a
-  `--out` file). The Rust binary spawns the adapter for `unit mutation --language typescript`;
-  `bin.ts` passes its `dist/` path to the binary as a `--ts-mutation-adapter` argument. Adds
-  `@stryker-mutator/api` as a devDependency (the engine's result types). The `bin` entry and package
-  exports are unchanged.
+  `--out` file). The Rust binary spawns the adapter (`dist/mutation/main.js`) for `unit mutation
+  --language typescript`; the launcher (`src/bin/index.ts`) passes its path to the binary as a
+  `--ts-mutation-adapter` argument. Adds `@stryker-mutator/api` as a devDependency (the engine's
+  result types). The `bin` entry now resolves to `dist/bin/index.js`; the package's `.` export is
+  unchanged.
 - **The TypeScript mutation engine ships with the package.** `@stryker-mutator/core` and
   `@stryker-mutator/vitest-runner` (`^9.6.0`) are declared as runtime dependencies, so installing
   testing-conventions brings them in and the adapter resolves them from the package's own tree; the
